Replace all the maps with a single fixed address one and use that for stack
Previously we were doing lots of mmaps to simulate mapping gecko's
omnijar. That caused the render thread's stack to change location,
which triggered the crash depending on that location.

This instead just does a single 1MB map at a fixed location, and uses
that for the render thread's stack. At certain addresses it crashes,
at others it does not.