ensure shaders are unique
This avoids the crash when using a single thread (or multiple threads
with locks).

The crash seems to occur when linking a program which contains a
shader with identical source to a previously compiled shader.

Even with unique shader sources, we still crash easily when compiling
shaders from multiple threads